+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Changelog
 ## [Unreleased]
+### Changed
+- The option `--loop-playlist` changed to `--loop` because it can be flagged in both `playlist` and `play` mode.
 
 ## [1.0.0] - 2025-02-17
-
 ### Added
 - The whole project